1.0.0-alpha.2

+++ build.sh
@@ -9,12 +9,17 @@ npx sort-package-json
 echo "3. building..."
 npm run build
 
-echo "4. build docs..."
-rm -r docs/
-npm run build-docs
-# Remove the first two lines from README file. The title is duplicated for some
-# reason.
-tail -n +3 ./docs/README.md > ./docs/temp && mv ./docs/temp ./docs/README.md
+if [ "$1" == 'skip-docs' ]
+then
+  echo "(skipping) 4. build docs..."
+else
+  echo "4. build docs..."
+  rm -r docs/
+  npm run build-docs
+  # Remove the first two lines from README file. The title is duplicated for
+  # some reason.
+  tail -n +3 ./docs/README.md > ./docs/temp && mv ./docs/temp ./docs/README.md
+fi
 
 echo "5. linting..."
 npm run lint
@@ -35,9 +40,9 @@ package_version=${package_version/,/} # remove `,`
 echo
 echo "--- BUILD COMPLETE ---"
 echo
-echo "1. update the package version if necessary: $package_version"
-echo "2. do not forget to run \"npm install\" if the package version is updated so that the lockfile is updated too"
-echo "3. create PR to main (from development), merge PR, and create a GitHub release"
-echo "4. switch to the main branch locally and run \"git pull\""
+echo "1. update the package version (currently: \"$package_version\") if necessary"
+echo "2. run \"npm install\" if the package version is updated to update the version in the lockfile"
+echo "3. create PR to master (from development), merge PR, and create a GitHub release"
+echo "4. switch to the master branch locally and run \"git pull\""
 echo "5. to publish the package run: \"npm publish --access public\""
-echo "6. switch back to development, merge main, and run \"git push\""
+echo "6. switch back to development, merge master, and run \"git pull\""
